Move X2AConfig to types.ts
