+++ workspaces/x2a/plugins/x2a-backend/config.d.ts
@@ -15,47 +15,6 @@
  */
 
 /**
- * X2A configuration structure
- * This is the runtime type extracted from app-config.yaml
- * Defaults for optional values are defined in src/services/constants.ts
- */
-export interface X2AConfig {
-  kubernetes: {
-    namespace: string;
-    image: string;
-    imageTag: string;
-    ttlSecondsAfterFinished: number;
-    resources: {
-      requests: {
-        cpu: string;
-        memory: string;
-      };
-      limits: {
-        cpu: string;
-        memory: string;
-      };
-    };
-  };
-  git?: {
-    author?: {
-      name: string;
-      email: string;
-    };
-  };
-  credentials: {
-    llm: Record<string, string>;
-    aap?: {
-      url: string;
-      orgName: string;
-      oauthToken?: string;
-      username?: string;
-      password?: string;
-      skipSSLVerification?: boolean;
-    };
-  };
-}
-
-/**
  * Root configuration schema for Backstage app-config.yaml
  * This Config interface extends Backstage's configuration structure to include X2A settings.
  * All X2A configuration values are optional with defaults defined in src/services/constants.ts
